MAIN : Force dynamic arrays size to prefix size if size is unknown. So if they end up empty the size will be correct.
Tests completed (copy as override , save and reload):

DIAL are OK except some float
QUST are OK

from A to L are OK except the same EXPL and the same light.
from M to Z one package is broken PatrolAndHunt, one region could not find its worldspace,
	plus a lot of float in NPC_ and Sounds

NAVI are still forbidden : the plugin will be corrupted if a NAVI is copied.

Update is OK.

Still checking CELL and WRLD
Hearthfires and Dawnguard are too big to apply the filter for cleaning apparently.

also listed one more unit in the dproj